+++ CHANGELOG.md
@@ -5,10 +5,9 @@ change log follows the conventions of
 
 ## [Unreleased]
 ### Added
-- Reagent interop for Menu
+- Reagent interop for Menu, Switch, Dialog
 - Reagent interop for Listbox, and recommendations for how to handle `:value`
   and `:on-change`.
-- Reagent interop for Switch
 - Support `:class` as a function. The function will receive the render props.
   Can be useful if the appearance of the component is the only thing that
   depends on the render props.